@@ -1,9 +1,9 @@
 {
   "name": "storycap-root",
   "private": true,
+  "type": "module",
   "scripts": {
     "clean": "lerna run clean",
-    "bootstrap": "lerna bootstrap",
     "doc:clihelp": "node scripts/cli-help.js",
     "doc:toc": "node scripts/add-toc.js",
     "doc": "yarn doc:clihelp && yarn doc:toc && prettier --write \"*.md\" \"packages/*/*.md\" --loglevel warn",
@@ -39,16 +39,12 @@
   },
   "homepage": "https://github.com/reg-viz/storycap#readme",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "8.46.2",
-    "@typescript-eslint/parser": "8.46.2",
     "copy-dir": "1.3.0",
-    "cpy": "8.1.2",
-    "cpy-cli": "5.0.0",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "8.10.0",
+    "eslint": "9.38.0",
+    "eslint-config-prettier": "10.1.8",
     "gh-pages": "6.3.0",
     "husky": "9.1.7",
-    "lerna": "4.0.0",
+    "lerna": "9.0.0",
     "markdown-toc": "1.2.0",
     "prettier": "3.6.2",
     "pretty-quick": "4.2.2",
@@ -57,6 +53,7 @@
     "reg-publish-s3-plugin": "0.14.4",
     "reg-suit": "0.14.5",
     "typescript": "5.9.3",
+    "typescript-eslint": "^8.46.2",
     "typescript-eslint-language-service": "5.0.5"
   },
   "workspaces": [
