@@ -1,3 +1,4 @@
-const { version } = require('../lerna.json');
+import lerna from '../lerna.json' with { type: 'json' };
+const { version } = lerna;
 const hit = version.match(/-(.+)\.\d+$/);
 console.log(hit ? 'next' : 'latest');
