@@ -235,7 +235,7 @@ export class CapturingBrowser extends StoryPreviewBrowser {
       } else {
         // Handle as Puppeteer device descriptor.
         const devices = await getDeviceDescriptors();
-        const hit = devices.find(d => d.name === opt.viewport);
+        const hit = devices.find(d => d.viewport === opt.viewport);
         if (!hit) {
           this.opt.logger.warn(
             `Skip screenshot for ${this.opt.logger.color.yellow(
@@ -378,11 +378,14 @@ export class CapturingBrowser extends StoryPreviewBrowser {
 
       if (forwardConsoleLogs) {
         switch (msg.type()) {
-          case 'warning':
+          case 'warn':
             logger.warn(niceMessage);
             break;
           case 'error':
             logger.error(niceMessage);
+            break;
+          case 'debug':
+            logger.debug(niceMessage);
             break;
           default:
             logger.log(niceMessage);
@@ -456,6 +459,7 @@ export class CapturingBrowser extends StoryPreviewBrowser {
       await this.waitForResources(mergedScreenshotOptions);
       await this.waitBrowserMetricsStable('postEmit');
 
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       await this.page.evaluate(() => new Promise(res => (window as any).requestIdleCallback(res, { timeout: 3000 })));
 
       // Get PNG image buffer
@@ -488,10 +492,11 @@ export class CapturingBrowser extends StoryPreviewBrowser {
       if (trace) {
         // Finish CPU trace.
         const traceBuffer = await this.page.tracing.stop();
-
-        // Calculate the suffix and save the trace to the file.
-        const suffix = variantKey.isDefault && defaultVariantSuffix ? [defaultVariantSuffix] : variantKey.keys;
-        await fileSystem.saveTrace(story.kind, story.story, suffix, traceBuffer);
+        if (traceBuffer) {
+          // Calculate the suffix and save the trace to the file.
+          const suffix = variantKey.isDefault && defaultVariantSuffix ? [defaultVariantSuffix] : variantKey.keys;
+          await fileSystem.saveTrace(story.kind, story.story, suffix, Buffer.from(traceBuffer));
+        }
       }
     }
   }
